Change event.id type from Integer to Long

+++ src/main/java/social/alone/server/event/controller/EventController.java
@@ -68,7 +68,8 @@ public class EventController extends BaseController {
 
   @DeleteMapping("/{id}")
   public ResponseEntity deleteEvent(
-          @PathVariable("id") Integer eventId) {
+          @PathVariable("id") Long eventId
+  ) {
 
     eventService.delete(eventId);
     return ResponseEntity.noContent().build();
